Minor updates. Directory name standardization, Image name standardization
