+++ OracleSOASuite/dockerfiles/12.2.1.2/container-scripts/createDomainAndStart.sh
