+++ OracleSOASuite/dockerfiles/12.2.1.2/container-scripts/updListenAddress.py
@@ -1,7 +1,5 @@
 #!/usr/bin/python
 #
-#
-#
 # Copyright (c) 2016-2017 Oracle and/or its affiliates. All rights reserved.
 #
 # Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
@@ -10,6 +8,7 @@
 # ==============================================
 #
 import sys
+
 #
 # Assigning values to variables
 # =============================
@@ -23,17 +22,17 @@ vol_name=sys.argv[1]
 manserver_host=sys.argv[2]
 server=sys.argv[3]
 exthost=sys.argv[4]
+
 #
 # Setting domain path
 # ===================
 domain_path  = domain_root + '/' + domain_name
-#
-#
+
 #
 # Read domain for updates
 # =======================
 readDomain(domain_path)
-#
+
 #
 # Set listen address
 # ==================
@@ -43,7 +42,7 @@ cmo.setListenAddress(manserver_host)
 cmo.setExternalDNSName(exthost)
 
 # Creating domain
-# ==============================
+# ===============
 updateDomain()
 closeDomain()
 exit()
