+++ OracleSOASuite/dockerfiles/12.2.1.2/soasuite.response
